requirements.txt
flask
google-generativeai
aiofiles